+++ src/jitlayers.h
@@ -5,7 +5,7 @@
 #include <llvm/IR/Module.h>
 #include <llvm/IR/Value.h>
 #include <llvm/IR/PassManager.h>
-#include "llvm/IR/LegacyPassManager.h"
+#include <llvm/IR/LegacyPassManager.h>
 
 #include <llvm/ExecutionEngine/Orc/IRCompileLayer.h>
 #include <llvm/ExecutionEngine/Orc/IRTransformLayer.h>
@@ -13,6 +13,10 @@
 
 #include <llvm/Target/TargetMachine.h>
 #include "julia_assert.h"
+#include "debug-registry.h"
+
+#include <stack>
+#include <queue>
 
 // As of LLVM 13, there are two runtime JIT linker implementations, the older
 // RuntimeDyld (used via orc::RTDyldObjectLinkingLayer) and the newer JITLink
@@ -48,9 +52,9 @@ using namespace llvm;
 
 extern "C" jl_cgparams_t jl_default_cgparams;
 
-void addTargetPasses(legacy::PassManagerBase *PM, TargetMachine *TM);
+void addTargetPasses(legacy::PassManagerBase *PM, const Triple &triple, TargetIRAnalysis analysis);
 void addOptimizationPasses(legacy::PassManagerBase *PM, int opt_level, bool lower_intrinsics=true, bool dump_native=false, bool external_use=false);
-void addMachinePasses(legacy::PassManagerBase *PM, TargetMachine *TM, int optlevel);
+void addMachinePasses(legacy::PassManagerBase *PM, int optlevel);
 void jl_finalize_module(orc::ThreadSafeModule  m);
 void jl_merge_module(orc::ThreadSafeModule &dest, orc::ThreadSafeModule src);
 GlobalVariable *jl_emit_RTLD_DEFAULT_var(Module *M);
@@ -59,6 +63,34 @@ DataLayout jl_create_datalayout(TargetMachine &TM);
 static inline bool imaging_default() {
     return jl_options.image_codegen || (jl_generating_output() && !jl_options.incremental);
 }
+
+struct jl_locked_stream {
+    JL_STREAM *stream = nullptr;
+    std::mutex mutex;
+
+    struct lock {
+        std::unique_lock<std::mutex> lck;
+        JL_STREAM *&stream;
+
+        lock(std::mutex &mutex, JL_STREAM *&stream) : lck(mutex), stream(stream) {}
+
+        JL_STREAM *&operator*() {
+            return stream;
+        }
+
+        explicit operator bool() {
+            return !!stream;
+        }
+
+        operator JL_STREAM *() {
+            return stream;
+        }
+    };
+
+    lock operator*() {
+        return lock(mutex, stream);
+    }
+};
 
 typedef struct _jl_llvm_functions_t {
     std::string functionObject;     // jlcall llvm Function name
@@ -169,7 +201,7 @@ static inline Constant *literal_static_pointer_val(const void *p, Type *T)
 #endif
 }
 
-static const inline char *name_from_method_instance(jl_method_instance_t *li)
+static const inline char *name_from_method_instance(jl_method_instance_t *li) JL_NOTSAFEPOINT
 {
     return jl_is_method(li->def.method) ? jl_symbol_name(li->def.method->name) : "top-level scope";
 }
@@ -193,37 +225,143 @@ public:
     typedef orc::IRCompileLayer CompileLayerT;
     typedef orc::IRTransformLayer OptimizeLayerT;
     typedef object::OwningBinary<object::ObjectFile> OwningObj;
-private:
-    struct OptimizerT {
-        OptimizerT(legacy::PassManager &PM, std::mutex &mutex, int optlevel) : optlevel(optlevel), PM(PM), mutex(mutex) {}
+    template
+    <typename ResourceT, size_t max = 0,
+        typename BackingT = std::stack<ResourceT,
+            std::conditional_t<max == 0,
+                SmallVector<ResourceT>,
+                SmallVector<ResourceT, max>
+            >
+        >
+    >
+    struct ResourcePool {
+        public:
+        ResourcePool(std::function<ResourceT()> creator) : creator(std::move(creator)), mutex(std::make_unique<WNMutex>()) {}
+        class OwningResource {
+            public:
+            OwningResource(ResourcePool &pool, ResourceT resource) : pool(pool), resource(std::move(resource)) {}
+            OwningResource(const OwningResource &) = delete;
+            OwningResource &operator=(const OwningResource &) = delete;
+            OwningResource(OwningResource &&) = default;
+            OwningResource &operator=(OwningResource &&) = default;
+            ~OwningResource() {
+                if (resource) pool.release(std::move(*resource));
+            }
+            ResourceT release() {
+                ResourceT res(std::move(*resource));
+                resource.reset();
+                return res;
+            }
+            void reset(ResourceT res) {
+                *resource = std::move(res);
+            }
+            ResourceT &operator*() {
+                return *resource;
+            }
+            ResourceT *operator->() {
+                return get();
+            }
+            ResourceT *get() {
+                return resource.getPointer();
+            }
+            const ResourceT &operator*() const {
+                return *resource;
+            }
+            const ResourceT *operator->() const {
+                return get();
+            }
+            const ResourceT *get() const {
+                return resource.getPointer();
+            }
+            explicit operator bool() const {
+                return resource;
+            }
+            private:
+            ResourcePool &pool;
+            llvm::Optional<ResourceT> resource;
+        };
 
-        OptimizerResultT operator()(orc::ThreadSafeModule M, orc::MaterializationResponsibility &R);
-    private:
-        int optlevel;
-        legacy::PassManager &PM;
-        std::mutex &mutex;
+        OwningResource operator*() {
+            return OwningResource(*this, acquire());
+        }
+
+        OwningResource get() {
+            return **this;
+        }
+
+        ResourceT acquire() {
+            std::unique_lock<std::mutex> lock(mutex->mutex);
+            if (!pool.empty()) {
+                return pop(pool);
+            }
+            if (!max || created < max) {
+                created++;
+                return creator();
+            }
+            mutex->empty.wait(lock, [&](){ return !pool.empty(); });
+            assert(!pool.empty() && "Expected resource pool to have a value!");
+            return pop(pool);
+        }
+        void release(ResourceT &&resource) {
+            std::lock_guard<std::mutex> lock(mutex->mutex);
+            pool.push(std::move(resource));
+            mutex->empty.notify_one();
+        }
+        private:
+        template<typename T, typename Container>
+        static ResourceT pop(std::queue<T, Container> &pool) {
+            ResourceT top = std::move(pool.front());
+            pool.pop();
+            return top;
+        }
+        template<typename PoolT>
+        static ResourceT pop(PoolT &pool) {
+            ResourceT top = std::move(pool.top());
+            pool.pop();
+            return top;
+        }
+        std::function<ResourceT()> creator;
+        size_t created = 0;
+        BackingT pool;
+        struct WNMutex {
+            std::mutex mutex;
+            std::condition_variable empty;
+        };
+
+        std::unique_ptr<WNMutex> mutex;
     };
-    // Custom object emission notification handler for the JuliaOJIT
-    template <typename ObjT, typename LoadResult>
-    void registerObject(const ObjT &Obj, const LoadResult &LO);
+    struct PipelineT {
+        PipelineT(orc::ObjectLayer &BaseLayer, TargetMachine &TM, int optlevel);
+        CompileLayerT CompileLayer;
+        OptimizeLayerT OptimizeLayer;
+    };
 
     struct OptSelLayerT : orc::IRLayer {
 
         template<size_t N>
-        OptSelLayerT(OptimizeLayerT (&optimizers)[N]) : orc::IRLayer(optimizers[0].getExecutionSession(), optimizers[0].getManglingOptions()), optimizers(optimizers), count(N) {
+        OptSelLayerT(const std::array<std::unique_ptr<PipelineT>, N> &optimizers)
+            : orc::IRLayer(optimizers[0]->OptimizeLayer.getExecutionSession(),
+                optimizers[0]->OptimizeLayer.getManglingOptions()),
+            optimizers(optimizers.data()),
+            count(N) {
             static_assert(N > 0, "Expected array with at least one optimizer!");
         }
 
         void emit(std::unique_ptr<orc::MaterializationResponsibility> R, orc::ThreadSafeModule TSM) override;
 
         private:
-        OptimizeLayerT *optimizers;
+        const std::unique_ptr<PipelineT> * const optimizers;
         size_t count;
     };
 
+private:
+    // Custom object emission notification handler for the JuliaOJIT
+    template <typename ObjT, typename LoadResult>
+    void registerObject(const ObjT &Obj, const LoadResult &LO);
+
 public:
 
-    JuliaOJIT(LLVMContext *Ctx);
+    JuliaOJIT();
 
     void enableJITDebuggingSupport();
 #ifndef JL_USE_JITLINK
@@ -239,44 +377,73 @@ public:
     uint64_t getGlobalValueAddress(StringRef Name);
     uint64_t getFunctionAddress(StringRef Name);
     StringRef getFunctionAtAddress(uint64_t Addr, jl_code_instance_t *codeinst);
-    orc::ThreadSafeContext &getContext();
+    auto getContext() {
+        return *ContextPool;
+    }
+    orc::ThreadSafeContext acquireContext() {
+        return ContextPool.acquire();
+    }
+    void releaseContext(orc::ThreadSafeContext &&ctx) {
+        ContextPool.release(std::move(ctx));
+    }
     const DataLayout& getDataLayout() const;
-    TargetMachine &getTargetMachine();
+
+    // TargetMachine pass-through methods
+    std::unique_ptr<TargetMachine> cloneTargetMachine() const;
     const Triple& getTargetTriple() const;
+    StringRef getTargetFeatureString() const;
+    StringRef getTargetCPU() const;
+    const TargetOptions &getTargetOptions() const;
+    const Target &getTarget() const;
+    TargetIRAnalysis getTargetIRAnalysis() const;
+
     size_t getTotalBytes() const;
+
+    JITDebugInfoRegistry &getDebugInfoRegistry() JL_NOTSAFEPOINT {
+        return DebugRegistry;
+    }
+
+    jl_locked_stream &get_dump_emitted_mi_name_stream() JL_NOTSAFEPOINT {
+        return dump_emitted_mi_name_stream;
+    }
+    jl_locked_stream &get_dump_compiles_stream() JL_NOTSAFEPOINT {
+        return dump_compiles_stream;
+    }
+    jl_locked_stream &get_dump_llvm_opt_stream() JL_NOTSAFEPOINT {
+        return dump_llvm_opt_stream;
+    }
 private:
     std::string getMangledName(StringRef Name);
     std::string getMangledName(const GlobalValue *GV);
     void shareStrings(Module &M);
 
-    std::unique_ptr<TargetMachine> TM;
-    DataLayout DL;
-    // Should be big enough that in the common case, The
-    // object fits in its entirety
-    legacy::PassManager PM0;  // per-optlevel pass managers
-    legacy::PassManager PM1;
-    legacy::PassManager PM2;
-    legacy::PassManager PM3;
-    std::mutex PM_mutexes[4];
-    std::unique_ptr<TargetMachine> TMs[4];
+    const std::unique_ptr<TargetMachine> TM;
+    const DataLayout DL;
 
-    orc::ThreadSafeContext TSCtx;
     orc::ExecutionSession ES;
     orc::JITDylib &GlobalJD;
     orc::JITDylib &JD;
 
+    JITDebugInfoRegistry DebugRegistry;
+
+    //Map and inc are guarded by RLST_mutex
+    std::mutex RLST_mutex{};
+    int RLST_inc = 0;
+    DenseMap<void*, std::string> ReverseLocalSymbolTable;
+
+    //Compilation streams
+    jl_locked_stream dump_emitted_mi_name_stream;
+    jl_locked_stream dump_compiles_stream;
+    jl_locked_stream dump_llvm_opt_stream;
+
+    ResourcePool<orc::ThreadSafeContext, 0, std::queue<orc::ThreadSafeContext>> ContextPool;
+
 #ifndef JL_USE_JITLINK
-    std::shared_ptr<RTDyldMemoryManager> MemMgr;
+    const std::shared_ptr<RTDyldMemoryManager> MemMgr;
 #endif
     ObjLayerT ObjectLayer;
-    CompileLayerT CompileLayer0;
-    CompileLayerT CompileLayer1;
-    CompileLayerT CompileLayer2;
-    CompileLayerT CompileLayer3;
-    OptimizeLayerT OptimizeLayers[4];
+    const std::array<std::unique_ptr<PipelineT>, 4> Pipelines;
     OptSelLayerT OptSelLayer;
-
-    DenseMap<void*, std::string> ReverseLocalSymbolTable;
 };
 extern JuliaOJIT *jl_ExecutionEngine;
 orc::ThreadSafeModule jl_create_llvm_module(StringRef name, orc::ThreadSafeContext ctx, bool imaging_mode, const DataLayout &DL = jl_ExecutionEngine->getDataLayout(), const Triple &triple = jl_ExecutionEngine->getTargetTriple());